fix int error

+++ train.py
@@ -187,12 +187,12 @@ def main(argv=None):
             (FLAGS.batch_size,FLAGS.seq_length-FLAGS.input_length-1))
         true_token = (random_flip < eta)
         #true_token = (random_flip < pow(base,itr))
-        ones = np.ones((FLAGS.img_width/FLAGS.patch_size,
-                        FLAGS.img_width/FLAGS.patch_size,
-                        FLAGS.patch_size**2*FLAGS.img_channel))
-        zeros = np.zeros((FLAGS.img_width/FLAGS.patch_size,
-                          FLAGS.img_width/FLAGS.patch_size,
-                          FLAGS.patch_size**2*FLAGS.img_channel))
+        ones = np.ones(int(FLAGS.img_width/FLAGS.patch_size),
+                        int(FLAGS.img_width/FLAGS.patch_size),
+                        int(FLAGS.patch_size**2*FLAGS.img_channel))
+        zeros = np.zeros((int(FLAGS.img_width/FLAGS.patch_size),
+                          int(FLAGS.img_width/FLAGS.patch_size),
+                          int(FLAGS.patch_size**2*FLAGS.img_channel)))
         mask_true = []
         for i in range(FLAGS.batch_size):
             for j in range(FLAGS.seq_length-FLAGS.input_length-1):
@@ -203,9 +203,9 @@ def main(argv=None):
         mask_true = np.array(mask_true)
         mask_true = np.reshape(mask_true, (FLAGS.batch_size,
                                            FLAGS.seq_length-FLAGS.input_length-1,
-                                           FLAGS.img_width/FLAGS.patch_size,
-                                           FLAGS.img_width/FLAGS.patch_size,
-                                           FLAGS.patch_size**2*FLAGS.img_channel))
+                                           int(FLAGS.img_width/FLAGS.patch_size),
+                                           int(FLAGS.img_width/FLAGS.patch_size),
+                                           int(FLAGS.patch_size**2*FLAGS.img_channel)))
         cost = model.train(ims, lr, mask_true)
         if FLAGS.reverse_input:
             ims_rev = ims[:,::-1]
